Speeding up test_core_profile_2d
core_profile_2d was setup to perform interpolation and return values
at (r,z). While this is good if we know beforehand what (r,z) will be
used, if one needs to use this function for arbitrary unknown values
of (r,z), then it wastes time in performing the interpolation
calculations for each call.

Interpolations.jl is setup such that callable interpolating functions
can be created and passed around. In this commit, we return a function
that can be called with (r,z) to return the interpolated values. This
is a factor of 10 or more faster when called repeatedly.