Updating dependency Interpolations.kl
This is to resolve the pinning of Interpolations.jl package to the
old 0.13.7 version as mentioned in
https://github.com/ProjectTorreyPines/SD4SOLPS.jl/issues/20
The NumericalIntegration.jl package is obsolete now, and thus has been
removed from our dependecy. We were only using it for calling
cumulative inegration on a known grid of points which is a simple
self contained function. So I copied the code of cumul_integrate
functiom from NumericalIntegration.jl
https://github.com/dextorious/NumericalIntegration.jl/blob/540b6c4bee089dfef7b9ae46e8ff188382e7c42e/src/NumericalIntegration.jl#L290
 (note this package has MIT license) and modified it to remove all
 macros which are considered potentially dangerous by Julia.
Now, we can use SD4SOLPS along with other packages that are using
new version of Interpolations.jl.