Using readg() with set_time & conformed to IMASDD
Using set_time for sample files where description is not in the
standard format.

Changes were made in the code to use ismissing in places to conform to
requirements of IMASDD.

In record_regular_mesh!(), the deepcopy of each element was replaced
with deepcopy of the entire grid subset to improve speed.

The test runtests.jl was updated to reflect the changes.

All tests pass now with following versions:

https://github.com/JuliaFusion/EFIT.jl/pull/8/commits/8580ca867807ebe512f186369236c566e8307f7c
https://github.com/ProjectTorreyPines/IMASDD.jl/commit/6a6968a0ed7dcd27120af03d6c9e6a24067e30b2
https://github.com/ProjectTorreyPines/GGDUtils.jl/commit/b11ad1581718a7bf4a50c2551b2a2a7b84a3394e
https://github.com/ProjectTorreyPines/SOLPS2IMAS.jl/commit/f843e6a64ccd597e0eb77ab4f953c922614d99e4

Note that EFIT version that was used has not been merged with its
master branch yet at the time of this test.

In conclusion, in future, OMAS can be removed now.