Using TPS interpolation to get midplane profile
For core profile extrapolation, the property needs to be evaluated at
the midplane surface. This is done by interpolating the property at the
midplane surface using the TPS interpolation method now. Earlier, it
was using KDTree method which was creating a very flat electron density
profile near the core boundary.